+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 __Enhancements__
 
-- New-SC365ExoReport now also includes hidden Microsoft outbound "Testmode" connectors
+- New-SC365ExoReport:
+  - now also includes hidden Microsoft outbound "Testmode" connectors
+  - Adds the logged on User as Report-Creator in the header
+  - Now has a transparent Logo
 - New-SC365Rules now supports -InternalSignature Parameter. This setting adds the two required rules to support this service. ONLY USE WHEN SERVICE is PROVISIONED, otherwise mailflow will fail.
 - Added Argument-Completers (automatically select correct values by pressing TAB after a parameter) for
   - New-SC365Connectors -MailDomain